717 - Reorganize - google-commons
