+++ server/libs/modules/components/google/google-calendar/build.gradle.kts
@@ -4,4 +4,5 @@ dependencies {
     implementation ("com.google.api-client:google-api-client:2.2.0")
     implementation ("com.google.oauth-client:google-oauth-client-jetty:1.34.1")
     implementation ("com.google.apis:google-api-services-calendar:v3-rev20231123-2.0.0")
+    api(project(":server:libs:modules:components:google:google-commons"))
 }
