+++ server/libs/modules/components/google/google-commons/build.gradle.kts
@@ -1,8 +1,8 @@
 version="1.0"
 
-
 dependencies {
     implementation ("com.google.api-client:google-api-client:2.0.0")
+    implementation ("com.google.apis:google-api-services-calendar:v3-rev20231123-2.0.0")
     implementation ("com.google.apis:google-api-services-drive:v3-rev20220815-2.0.0")
-    api(project(":server:libs:modules:components:google:google-commons"))
+    implementation("com.google.apis:google-api-services-gmail:v1-rev20220404-2.0.0")
 }
