+++ server/libs/modules/components/google/google-mail/build.gradle.kts
@@ -5,4 +5,5 @@ dependencies {
     implementation("com.google.oauth-client:google-oauth-client-jetty:1.34.1")
     implementation("com.google.apis:google-api-services-gmail:v1-rev20220404-2.0.0")
     implementation("org.eclipse.angus:angus-mail:2.0.2")
+    api(project(":server:libs:modules:components:google:google-commons"))
 }
