+++ server/libs/modules/components/google/google-mail/src/main/java/com/bytechef/component/google/mail/action/GoogleMailSearchEmailAction.java
@@ -36,7 +36,7 @@ import static com.bytechef.component.google.mail.constant.GoogleMailConstants.SE
 import com.bytechef.component.definition.ActionContext;
 import com.bytechef.component.definition.ComponentDSL.ModifiableActionDefinition;
 import com.bytechef.component.definition.Parameters;
-import com.bytechef.component.google.mail.util.GoogleMailUtils;
+import com.bytechef.google.commons.GoogleServices;
 import com.google.api.services.gmail.Gmail;
 import com.google.api.services.gmail.model.ListMessagesResponse;
 import java.io.IOException;
@@ -93,7 +93,7 @@ public class GoogleMailSearchEmailAction {
     public static ListMessagesResponse perform(
         Parameters inputParameters, Parameters connectionParameters, ActionContext actionContext) throws IOException {
 
-        Gmail service = GoogleMailUtils.getMail(connectionParameters);
+        Gmail service = GoogleServices.getMail(connectionParameters);
 
         return service.users()
             .messages()
