+++ settings.gradle.kts
@@ -168,6 +168,7 @@ include("server:libs:modules:components:example")
 include("server:libs:modules:components:file-storage")
 include("server:libs:modules:components:filesystem")
 include("server:libs:modules:components:google:google-calendar")
+include("server:libs:modules:components:google:google-commons")
 include("server:libs:modules:components:google:google-drive")
 include("server:libs:modules:components:google:google-mail")
 include("server:libs:modules:components:http-client")
